Keep tls_channel_options in Core_C
